requirements.txt
annotated-types==0.7.0
anyio==4.10.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
click-option-group==0.5.7
Deprecated==1.2.18
dotty-dict==1.3.1
filetype==1.2.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
google-genai==1.29.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.74
langchain-google-genai==2.1.9
langchain-text-splitters==0.3.9
langgraph==0.6.4
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.0
langsmith==0.4.13
lxml==6.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-gitlab==6.2.0
python-pptx==1.0.2
python-semantic-release==10.3.1
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
tenacity==9.1.2
tomlkit==0.13.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
websockets==15.0.1
wrapt==1.17.2
xlsxwriter==3.2.5
xxhash==3.5.0
zstandard==0.23.0
